requirements.txt
fastapi
uvicorn
gunicorn
pandas
numpy
scikit-learn
joblib
geopandas
shapely
pyproj
requests
python-dotenv
